feat: 팀 표시/숨기기, 모임 일정 모음 구현 완
