+++ src/components/FavorTeamSche/FavorTeamSche.tsx
@@ -65,7 +65,8 @@ export default function FavorTeamSche() {
           name: 'check222222222222',
           startDate: new Date('2024-01-11'),
           endDate: new Date('2024-01-14'),
-          explanation: 'Check that it should be appeared at FavorTeam scheBox. string이 칸을 넘어갈 경우 공백을 기준으로 줄바꿈이 이루어짐.',
+          explanation:
+            'Check that it should be appeared at FavorTeam scheBox. string이 칸을 넘어갈 경우 공백을 기준으로 줄바꿈이 이루어짐.',
         },
         {
           name: 'check3',
