+++ src/components/TeamShowBox/TeamShowBox.css
@@ -1,0 +1,6 @@
+.scroll::-webkit-scrollbar {
+  display: none;
+}
+.scroll {
+  -ms-overflow-style: none;
+}
